requirements.txt
google-cloud-vision==3.10.2
google-cloud==0.34.0
google-cloud-storage==3.2.0
firebase-admin==7.0.0
google-cloud-vision==3.10.2
google-cloud-pubsub==2.31.0
Flask==3.1.0
gunicorn